use floats just in case

+++ src/php-types/IntType.ts
@@ -2,11 +2,11 @@ import PhpType from '@/php-types/PhpType';
 
 export default class IntType extends PhpType {
     public getType(): string {
-        return 'int';
+        return 'float'; // safer?
     }
 
     public getDocblockContent(): string {
-        return 'int';
+        return 'float'; // safer?
     }
 
     public isDocblockRequired(): boolean {
